requirements.txt
streamlit
pandas
numpy
scipy
statsmodels
plotly
Pillow
os
pathlib